Clean up the scripts
