--- explore
@@ -1,4 +1,0 @@
-#!/usr/bin/env bash
-export DYLD_LIBRARY_PATH=~/.rustup/toolchains/nightly-2022-05-23-aarch64-apple-darwin/lib 
-export LD_LIBRARY_PATH=~/.rustup/toolchains/nightly-2022-05-23-x86_64-unknown-linux-gnu/lib 
-dfpp-explorer $@
